Extract directory_operation tests to it's own test file
